just changing the values along with fixing the return on double calculatePayDay function to return as a double rounded value

+++ src/Assignment3/HourlyWorker.java
@@ -45,7 +45,7 @@ public class HourlyWorker extends Employee {
     //But use of hourlyRate and hoursPerWeek
     public double calculatePayDay() {
 
-        return  Math.round((hourlyRate * hoursPerWeek) * 100) / 100;
+        return (double) Math.round((hourlyRate * hoursPerWeek) * 100) / 100;
     }
 
 }
